feat(cordova-webview-ad/ios): moved code from cordova/ios
